manager: add callable support to `subset_hook_caller`
The current `subset_hook_caller()` takes a static list of plugins to
exclude. pytest uses this functionality to exclude conftest plugins from
the wrong directory. The conftest business is very dynamic, which makes
the static nature of `subset_hook_caller()` very restricting, preventing
pytest for reusing subset callers for fear of a new conftest getting
loaded in the meantime.

Allow `subset_hook_caller` to also accept a callable; this would allow
pytest to pass a dynamic closure and avoid the invalidation issue.

I'll note that the above problem is mostly a weakness in how pytest
currently handles conftest loading, and should probably be fixed there
eventually. But the ability to pass a predicate instead of static set
seems like a useful generalization on its own.